Move test files into specs folder
