+++ test/specs/commands/encryptMessage.js
@@ -13,22 +13,22 @@
  * Removal or modification of this copyright notice is prohibited.
  *
  */
-import decryptMessageCommand from '../../src/commands/decryptMessage';
-import cryptoModule from '../../src/utils/cryptoModule';
-import * as input from '../../src/utils/input';
-import * as print from '../../src/utils/print';
+import encryptMessageCommand from '../../../src/commands/encryptMessage';
+import cryptoModule from '../../../src/utils/cryptoModule';
+import * as input from '../../../src/utils/input';
+import * as print from '../../../src/utils/print';
 import {
 	getCommands,
 	getRequiredArgs,
 	setUpVorpalWithCommand,
 } from './utils';
 
-describe('decrypt message command', () => {
-	const command = 'decrypt message';
+describe('encrypt message command', () => {
+	const command = 'encrypt message';
 	let vorpal;
 
 	beforeEach(() => {
-		vorpal = setUpVorpalWithCommand(decryptMessageCommand);
+		vorpal = setUpVorpalWithCommand(encryptMessageCommand);
 	});
 
 	afterEach(() => {
@@ -37,57 +37,60 @@ describe('decrypt message command', () => {
 
 	describe('setup', () => {
 		it('should be available', () => {
-			const decryptCommands = getCommands(vorpal, command);
-			(decryptCommands).should.have.length(1);
+			const encryptCommands = getCommands(vorpal, command);
+			(encryptCommands).should.have.length(1);
 		});
 
-		it('should require 2 inputs', () => {
+		it('should require 1 input', () => {
 			const requiredArgs = getRequiredArgs(vorpal, command);
-			(requiredArgs).should.have.length(2);
+			(requiredArgs).should.have.length(1);
 		});
 	});
 
 	describe('when executed', () => {
-		const data = 'Hello Lisker';
-		const encryptedData = '4728715ed4463a37d8e90720a27377f04a84911b95520c2582a8b6da';
-		const nonce = '682be05eeb73a794163b5584cac6b33769c2abd867459cae';
-		const passphrase = 'recipient secret';
-		// sender passphrase: 'sender secret'
-		const senderPublicKey = '38433137692948be1c05bbae686c9c850d3c8d9c52c1aebb4a7c1d5dd6d010d7';
+		const data = 'Some important message';
+		const multilineData = 'Some important message\nthat spans\nmultiple lines\n';
+		const passphrase = 'pass phrase';
+		const recipient = 'bba7e2e6a4639c431b68e31115a71ffefcb4e025a4d1656405dfdcd8384719e0';
 		const defaultPassphraseSource = `pass:${passphrase}`;
-		const commandWithSenderAndNonce = `${command} ${senderPublicKey} ${nonce}`;
-		const commandWithMessage = `${commandWithSenderAndNonce} ${encryptedData}`;
-		const commandWithPassphrase = `${commandWithSenderAndNonce} --passphrase "${defaultPassphraseSource}"`;
-		const cryptoDecryptReturnObject = { data };
+		const commandWithRecipient = `${command} ${recipient}`;
+		const commandWithMessage = `${commandWithRecipient} "${data}"`;
+		const commandWithPassphrase = `${commandWithRecipient} --passphrase "${defaultPassphraseSource}"`;
 
+		const nonce = '60ee6cbb5f9f0ee3736a6ffd20317f59ebfee2083e819909';
+		const encryptedMessage = '4ba04a1c568b66fe5f6e670295cd9945730013f4e3feb5ac0b4e3c';
+		const cryptoEncryptReturnObject = {
+			nonce,
+			encryptedMessage,
+		};
 		const defaultErrorMessage = 'Some error message.';
-		const wrappedErrorMessage = `Could not decrypt message: ${defaultErrorMessage}`;
+		const wrappedErrorMessage = `Could not encrypt message: ${defaultErrorMessage}`;
 
 		let stdInResult;
 		let getStdInStub;
 		let getPassphraseStub;
 		let getDataStub;
-		let decryptStub;
+		let encryptStub;
 		let printSpy;
 		let printResultStub;
 
 		beforeEach(() => {
 			getStdInStub = sandbox.stub(input, 'getStdIn').resolves({});
 			getPassphraseStub = sandbox.stub(input, 'getPassphrase').resolves(passphrase);
-			getDataStub = sandbox.stub(input, 'getData').resolves(encryptedData);
-			decryptStub = sandbox.stub(cryptoModule, 'decryptMessage').returns(cryptoDecryptReturnObject);
+			getDataStub = sandbox.stub(input, 'getData').resolves(data);
+			encryptStub = sandbox.stub(cryptoModule, 'encryptMessage').returns(cryptoEncryptReturnObject);
 			printSpy = sandbox.spy();
 			printResultStub = sandbox.stub(print, 'printResult').returns(printSpy);
 		});
 
 		describe('if no message source has been provided', () => {
 			beforeEach(() => {
-				return vorpal.exec(commandWithSenderAndNonce);
+				return vorpal.exec(commandWithRecipient);
 			});
 
-			it('should inform the user that the decryption was not successful', () => {
+			it('should inform the user that the encryption was not successful', () => {
 				(printResultStub.calledWithExactly(vorpal, {})).should.be.true();
-				(printSpy.calledWithExactly({ error: 'Could not decrypt message: No message was provided.' })).should.be.true();
+				(printSpy.calledWithExactly({ error: 'Could not encrypt message: No message was provided.' })).should.be.true();
 			});
 		});
 
@@ -97,19 +100,19 @@ describe('decrypt message command', () => {
 				return vorpal.exec(commandWithMessage);
 			});
 
-			it('should inform the user the decryption was not successful', () => {
+			it('should inform the user the encryption was not successful', () => {
 				(printResultStub.calledWithExactly(vorpal, {})).should.be.true();
 				(printSpy.calledWithExactly({ error: wrappedErrorMessage })).should.be.true();
 			});
 		});
 
-		describe('if an error occurs during decryption', () => {
+		describe('if an error occurs during encryption', () => {
 			beforeEach(() => {
-				decryptStub.rejects(new Error(defaultErrorMessage));
+				encryptStub.rejects(new Error(defaultErrorMessage));
 				return vorpal.exec(commandWithMessage);
 			});
 
-			it('should inform the user the decryption was not successful', () => {
+			it('should inform the user the encryption was not successful', () => {
 				(printResultStub.calledWithExactly(vorpal, {})).should.be.true();
 				(printSpy.calledWithExactly({ error: wrappedErrorMessage })).should.be.true();
 			});
@@ -122,7 +125,7 @@ describe('decrypt message command', () => {
 					return vorpal.exec(commandWithMessage);
 				});
 
-				it('should inform the user the decryption was not successful', () => {
+				it('should inform the user the encryption was not successful', () => {
 					(printResultStub.calledWithExactly(vorpal, {})).should.be.true();
 					(printSpy.calledWithExactly({ error: wrappedErrorMessage })).should.be.true();
 				});
@@ -142,23 +145,25 @@ describe('decrypt message command', () => {
 				});
 
 				it('should call the input util getPassphrase with the correct parameters', () => {
-					(getPassphraseStub.calledWithExactly(vorpal, undefined, undefined))
+					(getPassphraseStub.calledWithExactly(
+						vorpal, undefined, undefined, { shouldRepeat: true },
+					))
 						.should.be.true();
 				});
 
 				it('should call the input util getData with the correct parameters', () => {
-					(getDataStub.calledWithExactly(encryptedData, undefined, undefined))
+					(getDataStub.calledWithExactly(data, undefined, undefined))
 						.should.be.true();
 				});
 
-				it('should call the crypto module decrypt method with correct parameters', () => {
-					(decryptStub.calledWithExactly(encryptedData, nonce, passphrase, senderPublicKey))
+				it('should call the crypto module encrypt method with correct parameters', () => {
+					(encryptStub.calledWithExactly(data, passphrase, recipient))
 						.should.be.true();
 				});
 
 				it('should print the result', () => {
 					(printResultStub.calledWithExactly(vorpal, {})).should.be.true();
-					(printSpy.calledWithExactly(cryptoDecryptReturnObject)).should.be.true();
+					(printSpy.calledWithExactly(cryptoEncryptReturnObject)).should.be.true();
 				});
 			});
 
@@ -179,24 +184,25 @@ describe('decrypt message command', () => {
 				});
 
 				it('should call the input util getPassphrase method with correct parameters', () => {
-					(getPassphraseStub.calledWithExactly(vorpal, passphraseSource, undefined))
+					(getPassphraseStub.calledWithExactly(
+						vorpal, passphraseSource, undefined, { shouldRepeat: true },
+					))
 						.should.be.true();
 				});
 
 				it('should call the input util getData with the correct parameters', () => {
-					(getDataStub.calledWithExactly(encryptedData, undefined, undefined))
+					(getDataStub.calledWithExactly(data, undefined, undefined))
 						.should.be.true();
 				});
 
-				it('should call the crypto module decrypt method with correct parameters', () => {
-					(decryptStub.calledWithExactly(encryptedData, nonce, passphrase, senderPublicKey))
-						.should.be.true();
+				it('should call the crypto module encrypt method with correct parameters', () => {
+					(encryptStub.calledWithExactly(data, passphrase, recipient)).should.be.true();
 				});
 
 				it('should print the result', () => {
 					(printResultStub.calledWithExactly(vorpal, { passphrase: passphraseSource }))
 						.should.be.true();
-					(printSpy.calledWithExactly(cryptoDecryptReturnObject)).should.be.true();
+					(printSpy.calledWithExactly(cryptoEncryptReturnObject)).should.be.true();
 				});
 			});
 
@@ -218,24 +224,25 @@ describe('decrypt message command', () => {
 				});
 
 				it('should call the input util getPassphrase with correct parameters', () => {
-					(getPassphraseStub.calledWithExactly(vorpal, passphraseSource, undefined))
+					(getPassphraseStub.calledWithExactly(
+						vorpal, passphraseSource, undefined, { shouldRepeat: true },
+					))
 						.should.be.true();
 				});
 
 				it('should call the input util getData with the correct parameters', () => {
-					(getDataStub.calledWithExactly(encryptedData, undefined, undefined))
+					(getDataStub.calledWithExactly(data, undefined, undefined))
 						.should.be.true();
 				});
 
-				it('should call the crypto module decrypt method with correct parameters', () => {
-					(decryptStub.calledWithExactly(encryptedData, nonce, passphrase, senderPublicKey))
-						.should.be.true();
+				it('should call the crypto module encrypt method with correct parameters', () => {
+					(encryptStub.calledWithExactly(data, passphrase, recipient)).should.be.true();
 				});
 
 				it('should print the result', () => {
 					(printResultStub.calledWithExactly(vorpal, { passphrase: passphraseSource }))
 						.should.be.true();
-					(printSpy.calledWithExactly(cryptoDecryptReturnObject)).should.be.true();
+					(printSpy.calledWithExactly(cryptoEncryptReturnObject)).should.be.true();
 				});
 			});
 
@@ -256,24 +263,26 @@ describe('decrypt message command', () => {
 				});
 
 				it('should call the input util getPassphrase method with correct parameters', () => {
-					(getPassphraseStub.calledWithExactly(vorpal, passphraseSource, undefined))
+					(getPassphraseStub.calledWithExactly(
+						vorpal, passphraseSource, undefined, { shouldRepeat: true },
+					))
 						.should.be.true();
 				});
 
 				it('should call the input util getData with the correct parameters', () => {
-					(getDataStub.calledWithExactly(encryptedData, undefined, undefined))
+					(getDataStub.calledWithExactly(data, undefined, undefined))
 						.should.be.true();
 				});
 
-				it('should call the crypto module decrypt method with correct parameters', () => {
-					(decryptStub.calledWithExactly(encryptedData, nonce, passphrase, senderPublicKey))
+				it('should call the crypto module encrypt method with correct parameters', () => {
+					(encryptStub.calledWithExactly(data, passphrase, recipient))
 						.should.be.true();
 				});
 
 				it('should print the result', () => {
 					(printResultStub.calledWithExactly(vorpal, { passphrase: passphraseSource }))
 						.should.be.true();
-					(printSpy.calledWithExactly(cryptoDecryptReturnObject)).should.be.true();
+					(printSpy.calledWithExactly(cryptoEncryptReturnObject)).should.be.true();
 				});
 			});
 
@@ -296,45 +305,48 @@ describe('decrypt message command', () => {
 				});
 
 				it('should call the input util getPassphrase with the correct parameters', () => {
-					(getPassphraseStub.calledWithExactly(vorpal, passphraseSource, passphrase))
+					(getPassphraseStub.calledWithExactly(
+						vorpal, passphraseSource, passphrase, { shouldRepeat: true },
+					))
 						.should.be.true();
 				});
 
 				it('should call the input util getData with the correct parameters', () => {
-					(getDataStub.calledWithExactly(encryptedData, undefined, undefined))
+					(getDataStub.calledWithExactly(data, undefined, undefined))
 						.should.be.true();
 				});
 
-				it('should call the crypto module decrypt method with correct parameters', () => {
-					(decryptStub.calledWithExactly(encryptedData, nonce, passphrase, senderPublicKey))
+				it('should call the crypto module encrypt method with correct parameters', () => {
+					(encryptStub.calledWithExactly(data, passphrase, recipient))
 						.should.be.true();
 				});
 
 				it('should print the result', () => {
 					(printResultStub.calledWithExactly(vorpal, { passphrase: passphraseSource }))
 						.should.be.true();
-					(printSpy.calledWithExactly(cryptoDecryptReturnObject)).should.be.true();
+					(printSpy.calledWithExactly(cryptoEncryptReturnObject)).should.be.true();
 				});
 			});
 		});
 
-		describe('encrypted data', () => {
+		describe('data', () => {
 			describe('if the data cannot be retrieved', () => {
 				beforeEach(() => {
 					getDataStub.rejects(new Error(defaultErrorMessage));
 					return vorpal.exec(commandWithMessage);
 				});
 
-				it('should inform the user the decryption was not successful', () => {
+				it('should inform the user the encryption was not successful', () => {
 					(printResultStub.calledWithExactly(vorpal, {})).should.be.true();
 					(printSpy.calledWithExactly({ error: wrappedErrorMessage })).should.be.true();
 				});
 			});
 
 			describe('with data passed as a command line argument', () => {
-				const dataPlainTextCommand = `${commandWithPassphrase} "${encryptedData}"`;
+				const dataPlainTextCommand = `${commandWithPassphrase} "${data}"`;
 
 				beforeEach(() => {
+					getDataStub.resolves(data);
 					return vorpal.exec(dataPlainTextCommand);
 				});
 
@@ -347,24 +359,26 @@ describe('decrypt message command', () => {
 				});
 
 				it('should call the input util getPassphrase with the correct parameters', () => {
-					(getPassphraseStub.calledWithExactly(vorpal, defaultPassphraseSource, undefined))
+					(getPassphraseStub.calledWithExactly(
+						vorpal, defaultPassphraseSource, undefined, { shouldRepeat: true },
+					))
 						.should.be.true();
 				});
 
 				it('should call the input util getData with the correct parameters', () => {
-					(getDataStub.calledWithExactly(encryptedData, undefined, undefined))
+					(getDataStub.calledWithExactly(data, undefined, undefined))
 						.should.be.true();
 				});
 
-				it('should call the crypto module decrypt method with correct parameters', () => {
-					(decryptStub.calledWithExactly(encryptedData, nonce, passphrase, senderPublicKey))
+				it('should call the crypto module encrypt method with correct parameters', () => {
+					(encryptStub.calledWithExactly(data, passphrase, recipient))
 						.should.be.true();
 				});
 
 				it('should print the result', () => {
 					(printResultStub.calledWithExactly(vorpal, { passphrase: defaultPassphraseSource }))
 						.should.be.true();
-					(printSpy.calledWithExactly(cryptoDecryptReturnObject)).should.be.true();
+					(printSpy.calledWithExactly(cryptoEncryptReturnObject)).should.be.true();
 				});
 			});
 
@@ -372,6 +386,7 @@ describe('decrypt message command', () => {
 				const dataFileCommand = `${commandWithPassphrase} --message file:/path/to/message.txt`;
 
 				beforeEach(() => {
+					getDataStub.resolves(multilineData);
 					return vorpal.exec(dataFileCommand);
 				});
 
@@ -384,7 +399,9 @@ describe('decrypt message command', () => {
 				});
 
 				it('should call the input util getPassphrase with the correct parameters', () => {
-					(getPassphraseStub.calledWithExactly(vorpal, defaultPassphraseSource, undefined))
+					(getPassphraseStub.calledWithExactly(
+						vorpal, defaultPassphraseSource, undefined, { shouldRepeat: true },
+					))
 						.should.be.true();
 				});
 
@@ -393,15 +410,15 @@ describe('decrypt message command', () => {
 						.should.be.true();
 				});
 
-				it('should call the crypto module decrypt method with correct parameters', () => {
-					(decryptStub.calledWithExactly(encryptedData, nonce, passphrase, senderPublicKey))
+				it('should call the crypto module encrypt method with correct parameters', () => {
+					(encryptStub.calledWithExactly(multilineData, passphrase, recipient))
 						.should.be.true();
 				});
 
 				it('should print the result', () => {
 					(printResultStub.calledWithExactly(vorpal, { passphrase: defaultPassphraseSource, message: 'file:/path/to/message.txt' }))
 						.should.be.true();
-					(printSpy.calledWithExactly(cryptoDecryptReturnObject)).should.be.true();
+					(printSpy.calledWithExactly(cryptoEncryptReturnObject)).should.be.true();
 				});
 			});
 
@@ -409,8 +426,9 @@ describe('decrypt message command', () => {
 				const dataStdInCommand = `${commandWithPassphrase} --message stdin`;
 
 				beforeEach(() => {
-					stdInResult = { data: encryptedData };
+					stdInResult = { data: multilineData };
 					getStdInStub.resolves(stdInResult);
+					getDataStub.resolves(multilineData);
 					return vorpal.exec(dataStdInCommand);
 				});
 
@@ -423,36 +441,38 @@ describe('decrypt message command', () => {
 				});
 
 				it('should call the input util getPassphrase with the correct parameters', () => {
-					(getPassphraseStub.calledWithExactly(vorpal, defaultPassphraseSource, undefined))
+					(getPassphraseStub.calledWithExactly(
+						vorpal, defaultPassphraseSource, undefined, { shouldRepeat: true },
+					))
 						.should.be.true();
 				});
 
 				it('should call the input util getData with the correct parameters', () => {
-					(getDataStub.calledWithExactly(undefined, 'stdin', encryptedData))
+					(getDataStub.calledWithExactly(undefined, 'stdin', multilineData))
 						.should.be.true();
 				});
 
-				it('should call the crypto module decrypt method with correct parameters', () => {
-					(decryptStub.calledWithExactly(encryptedData, nonce, passphrase, senderPublicKey))
+				it('should call the crypto module encrypt method with correct parameters', () => {
+					(encryptStub.calledWithExactly(multilineData, passphrase, recipient))
 						.should.be.true();
 				});
 
 				it('should print the result', () => {
 					(printResultStub.calledWithExactly(vorpal, { passphrase: defaultPassphraseSource, message: 'stdin' }))
 						.should.be.true();
-					(printSpy.calledWithExactly(cryptoDecryptReturnObject)).should.be.true();
+					(printSpy.calledWithExactly(cryptoEncryptReturnObject)).should.be.true();
 				});
 			});
 		});
 
-		describe('with passphrase and encrypted data passed via stdin', () => {
-			const passphraseAndMessageStdInCommand = `${command} ${senderPublicKey} ${nonce} --message stdin --passphrase stdin`;
+		describe('with passphrase and data passed via stdin', () => {
+			const passphraseAndDataStdInCommand = `${command} ${recipient} --message stdin --passphrase stdin`;
 
 			beforeEach(() => {
-				stdInResult = { passphrase, data: encryptedData };
+				stdInResult = { passphrase, data: multilineData };
 				getStdInStub.resolves(stdInResult);
-				getDataStub.resolves(encryptedData);
-				return vorpal.exec(passphraseAndMessageStdInCommand);
+				getDataStub.resolves(multilineData);
+				return vorpal.exec(passphraseAndDataStdInCommand);
 			});
 
 			it('should call the input util getStdIn with the correct parameters', () => {
@@ -464,24 +484,24 @@ describe('decrypt message command', () => {
 			});
 
 			it('should call the input util getPassphrase with the correct parameters', () => {
-				(getPassphraseStub.calledWithExactly(vorpal, 'stdin', passphrase))
+				(getPassphraseStub.calledWithExactly(vorpal, 'stdin', passphrase, { shouldRepeat: true }))
 					.should.be.true();
 			});
 
 			it('should call the input util getData with the correct parameters', () => {
-				(getDataStub.calledWithExactly(undefined, 'stdin', encryptedData))
+				(getDataStub.calledWithExactly(undefined, 'stdin', multilineData))
 					.should.be.true();
 			});
 
 			it('should call the crypto module encrypt method with correct parameters', () => {
-				(decryptStub.calledWithExactly(encryptedData, nonce, passphrase, senderPublicKey))
+				(encryptStub.calledWithExactly(multilineData, passphrase, recipient))
 					.should.be.true();
 			});
 
 			it('should print the result', () => {
 				(printResultStub.calledWithExactly(vorpal, { passphrase: 'stdin', message: 'stdin' }))
 					.should.be.true();
-				(printSpy.calledWithExactly(cryptoDecryptReturnObject)).should.be.true();
+				(printSpy.calledWithExactly(cryptoEncryptReturnObject)).should.be.true();
 			});
 		});
 	});
